modified cli and index
